@@ -11,54 +11,51 @@
     "build": "backstage-cli package build",
     "clean": "backstage-cli package clean",
     "test": "backstage-cli package test",
-    "lint": "backstage-cli package lint",
-    "test:e2e": "cross-env PORT=3001 start-server-and-test start http://localhost:3001 cy:dev",
-    "test:e2e:ci": "cross-env PORT=3001 start-server-and-test start http://localhost:3001 cy:run",
-    "cy:dev": "cypress open",
-    "cy:run": "cypress run --browser chrome"
+    "lint": "backstage-cli package lint"
   },
   "dependencies": {
-    "@backstage/app-defaults": "^1.4.2",
-    "@backstage/catalog-model": "^1.4.1",
-    "@backstage/cli": "^0.22.12",
-    "@backstage/core-app-api": "^1.9.1",
-    "@backstage/core-components": "^0.13.4",
-    "@backstage/core-plugin-api": "^1.5.3",
-    "@backstage/integration-react": "^1.1.18",
-    "@backstage/plugin-api-docs": "^0.9.10",
-    "@backstage/plugin-catalog": "^1.12.4",
-    "@backstage/plugin-catalog-common": "^1.0.15",
-    "@backstage/plugin-catalog-graph": "^0.2.35",
-    "@backstage/plugin-catalog-import": "^0.9.13",
-    "@backstage/plugin-catalog-react": "^1.8.3",
-    "@backstage/plugin-github-actions": "^0.6.4",
-    "@backstage/plugin-org": "^0.6.13",
-    "@backstage/plugin-permission-react": "^0.4.14",
-    "@backstage/plugin-scaffolder": "^1.14.4",
-    "@backstage/plugin-search": "^1.3.6",
-    "@backstage/plugin-search-react": "^1.6.4",
-    "@backstage/plugin-tech-radar": "^0.6.7",
-    "@backstage/plugin-techdocs": "^1.6.8",
-    "@backstage/plugin-techdocs-module-addons-contrib": "^1.0.18",
-    "@backstage/plugin-techdocs-react": "^1.1.9",
-    "@backstage/plugin-user-settings": "^0.7.8",
-    "@backstage/theme": "^0.4.1",
+    "@backstage/app-defaults": "^1.4.8-next.1",
+    "@backstage/catalog-model": "^1.4.4-next.0",
+    "@backstage/cli": "^0.25.2-next.1",
+    "@backstage/core-app-api": "^1.11.4-next.0",
+    "@backstage/core-components": "^0.14.0-next.0",
+    "@backstage/core-plugin-api": "^1.8.3-next.0",
+    "@backstage/integration-react": "^1.1.24-next.0",
+    "@backstage/plugin-api-docs": "^0.10.4-next.1",
+    "@backstage/plugin-catalog": "^1.17.0-next.1",
+    "@backstage/plugin-catalog-common": "^1.0.21-next.0",
+    "@backstage/plugin-catalog-graph": "^0.3.4-next.1",
+    "@backstage/plugin-catalog-import": "^0.10.6-next.1",
+    "@backstage/plugin-catalog-react": "^1.9.4-next.1",
+    "@backstage/plugin-github-actions": "^0.6.11-next.1",
+    "@backstage/plugin-org": "^0.6.20-next.1",
+    "@backstage/plugin-permission-react": "^0.4.20-next.0",
+    "@backstage/plugin-scaffolder": "^1.18.0-next.1",
+    "@backstage/plugin-search": "^1.4.6-next.1",
+    "@backstage/plugin-search-react": "^1.7.6-next.1",
+    "@backstage/plugin-tech-radar": "^0.6.13-next.1",
+    "@backstage/plugin-techdocs": "^1.10.0-next.1",
+    "@backstage/plugin-techdocs-module-addons-contrib": "^1.1.5-next.1",
+    "@backstage/plugin-techdocs-react": "^1.1.16-next.0",
+    "@backstage/plugin-user-settings": "^0.8.1-next.1",
+    "@backstage/theme": "^0.5.0",
     "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.9.1",
-    "@spotify/backstage-plugin-soundcheck": "^0.8.1",
+    "@spotify/backstage-plugin-soundcheck": "^0.11.1",
     "git-url-parse": "^13.1.0",
     "history": "^5.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.2",
+    "react-dom": "^18.0.2",
     "react-router": "^6.3.0",
     "react-router-dom": "^6.3.0",
     "react-use": "^17.2.4"
   },
   "devDependencies": {
-    "@backstage/test-utils": "^1.4.2",
-    "@testing-library/dom": "^8.0.0",
-    "@testing-library/jest-dom": "^5.10.1",
-    "@testing-library/react": "^12.1.3",
+    "@backstage/test-utils": "^1.5.0-next.1",
+    "@playwright/test": "^1.32.3",
+    "@testing-library/jest-dom": "^6.0.0",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/dom": "^9.0.0",
     "@testing-library/user-event": "^14.0.0",
     "@types/node": "^16.11.26",
     "@types/react-dom": "*",
